+++ patient-generated-data-tests/src/test/java/gov/va/api/health/patientgenerateddata/tests/QuestionnaireResponseWritesIT.java
@@ -1,0 +1,51 @@
+package gov.va.api.health.patientgenerateddata.tests;
+
+import static gov.va.api.health.patientgenerateddata.tests.RequestUtils.doGet;
+import static gov.va.api.health.patientgenerateddata.tests.RequestUtils.doPut;
+import static gov.va.api.health.patientgenerateddata.tests.RequestUtils.serializePayload;
+import static gov.va.api.health.patientgenerateddata.tests.SystemDefinitions.systemDefinition;
+import static gov.va.api.health.sentinel.EnvironmentAssumptions.assumeEnvironmentIn;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import gov.va.api.health.r4.api.resources.QuestionnaireResponse;
+import gov.va.api.health.r4.api.resources.QuestionnaireResponse.Status;
+import gov.va.api.health.sentinel.Environment;
+import gov.va.api.health.sentinel.ExpectedResponse;
+import java.time.Instant;
+import java.time.temporal.ChronoField;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+public class QuestionnaireResponseWritesIT {
+
+  @BeforeAll
+  static void assumeEnvironment() {
+    // Anything but LAB
+    assumeEnvironmentIn(
+        Environment.LOCAL, Environment.QA, Environment.STAGING, Environment.STAGING_LAB);
+    loadInitialResource();
+  }
+
+  private static void loadInitialResource() {
+    var id = systemDefinition().ids().questionnaireResponse();
+    QuestionnaireResponse qr = questionnaireResponse(id);
+    RequestUtils.doPut(
+        "QuestionnaireResponse/" + id, serializePayload(qr), "load initial resource", true);
+  }
+
+  static QuestionnaireResponse questionnaireResponse(String id) {
+    return QuestionnaireResponse.builder().id(id).status(Status.completed).build();
+  }
+
+  @Test
+  void update_authored() {
+    var id = systemDefinition().ids().questionnaireResponse();
+    Instant now = Instant.now().with(ChronoField.NANO_OF_SECOND, 0);
+    QuestionnaireResponse qr = questionnaireResponse(id).authored(now.toString());
+    doPut("QuestionnaireResponse/" + id, serializePayload(qr), "update authored date", 200);
+    ExpectedResponse persistedResponse =
+        doGet("application/json", "QuestionnaireResponse/" + id, 200);
+    QuestionnaireResponse persisted = persistedResponse.response().as(QuestionnaireResponse.class);
+    assertThat(persisted.authored()).isEqualTo(now.toString());
+  }
+}
